Mobile Robot
The importance of autonomous robots are increasing day by day in parallel with developing technology and the changing needs at industry and people. In the aim of saving time and energy, sensitivity in industry and gaining advantage of accuracy and rapidness tendency of autonomous robots has shown a great improve recently. Nowadays, without any human needs autonomous robots, wireless charging is generated for the movement and process continuity. That is how robots can be isolated from human hand. Wireless charging has a main problem that is inefficiency, but we are going to foreground wireless charging’s advantages. The most important and remarkable property is that the robot can charge itself on its own. Circuit that designed to charge a 5Volt battery, after connecting to the source with an adapter, wireless power transmission begins. As a result of this transmission between the receiver and the transmitter, the command to activate the robot is given when the battery level is %100. The robot starts its random circulation after it completed its charging process. The aim of the robot is to go to charging station when its charge level drops below the specified charge level. The process of the searching for the charging station will act upon the transmitting signals of the charging station. The direction required for to arrive at the charging station will be determined through the IR receiver circuit on 3 different sides of robot. When the robot arrives at the station, charging is carried out in the form of wireless energy transmission. The process starts by regulating the AC voltage from the power supply and obtaining a voltage of 5 V DC. Through coils adjusted to the required energy level, energy transfer from transmitter to receiver is provided. This is how the robot charged wirelessly. The purpose of the robot is designated to be able to observe the environmental data in a closed environment and to be aware of unsafe events (gas leakage, smoke, fire, etc.). Therefore, an application that enables remote control of the robot has been created, and the data has been accessed through the web server, thus adding unique values to the project.
Keywords: Wireless charge, autonomous robot, direction finding